[!] Fixed MSVC warnings [!] Optimization for Sprite2D, only update the counter individually when next command doesn't match [!] Fixed and optimized guNormalize_Rugrats (Still can't find any game that uses this errg)
git-svn-id: https://subversion.assembla.com/svn/Daedalusx64/trunk@1599 42e9bfbe-799a-4a2d-bad1-236e862a387a

+++ Source/Core/R4300.cpp
@@ -3391,7 +3391,7 @@ static void R4300_CALL_TYPE R4300_Cop1_D_CVT_S( R4300_CALL_SIGNATURE )
 
 	SET_ROUND_MODE( gRoundingMode );		//XXXX Is this needed?
 
-	StoreFPR_Single( op_code.fd, fX );
+	StoreFPR_Single( op_code.fd, (f32)fX );
 }
 
 // Convert f64 to word...
