+++ Source/SysOSX/HLEGraphics/DisplayListDebugger.cpp
@@ -268,7 +268,7 @@ void DLDebugger_ProcessDebugTask()
 				glReadPixels(0, 0, width, height, GL_RGBA, GL_UNSIGNED_BYTE, pixels);
 
 				// NB, pass a negative pitch, to render the screenshot the right way up.
-				s32 pitch = -(width * 4);
+				s32 pitch = -static_cast<s32>(width * 4);
 
 				PngSaveImage(connection, pixels, NULL, TexFmt_8888, pitch, width, height, false);
